Completed DAOs

+++ GestioneSegreteriaLezione/src/main/java/it/unical/mat/webcomp21/persistence/dao/ScuolaDAO.java
@@ -7,7 +7,7 @@ import it.unical.mat.webcomp21.model.Scuola;
 
 public interface ScuolaDAO {
 	public void save(Scuola scuola);  // Create
-	public Scuola findByPrimaryKey(int id);     // Retrieve
+	public Scuola findByPrimaryKey(Long id);     // Retrieve
 	public List<Scuola> findAll();       
 	public void update(Scuola scuola); //Update
 	public void delete(Scuola scuola); //Delete
